+++ Package.swift
@@ -35,11 +35,11 @@ let package = Package(
         // Targets are the basic building blocks of a package. A target can define a module or a test suite.
         // Targets can depend on other targets in this package, and on products in packages which this package depends on.
         .target(
-            name: "STLR",
-            dependencies: ["OysterKit"]),
-        .target(
             name: "OysterKit",
             dependencies: []),
+        .target(
+            name: "STLR",
+            dependencies: ["OysterKit"]),
         .target(
             name: "stlr-cli",
             dependencies: ["OysterKit","STLR"]),
