+++ Sources/STLR/Optimisers/CharacterSetOnlyChoiceOptimizer.swift
