+++ Sources/STLR/Optimisers/InlineIdentifierOptimizer.swift
