--- Sources/STLR/deleteme.swift
@@ -1,3 +1,0 @@
-public class Wibble{
-	let a : Int = 0
-}
